compiler/pyproject.toml
[tool.poetry]
name = "compiler"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = ">=3.7.10,<3.9"
click = ">=7,<8"
PyYAML = "^5.4.1"
tfx = ">=1.0.0"
kfp = "^1.8.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
